Renamed robot to constants
