+++ src/constants/robot_motor.cpp
@@ -1,13 +1,13 @@
-#include "robot_motor.h"
+#include "constants/motor_constants.h"
 
-namespace robot {
-pros::motor_gearset_e_t kDrivetrainGearset = pros::E_MOTOR_GEAR_BLUE;
-std::vector<bool> kDrivetrainReverse = {false, true};
+namespace constants {
+const pros::motor_gearset_e_t kDrivetrainGearset = pros::E_MOTOR_GEAR_BLUE;
+const std::vector<bool> kDrivetrainReverse = {false, true};
 // {top motor, bottom motor}
-std::vector<int> kFrontRightMotorPorts = {1, 2};
-std::vector<int> kBackRightMotorPorts = {4, 5};
-std::vector<int> kBackLeftMotorPorts = {6, 7};
-std::vector<int> kFrontLeftMotorPorts = {3, 4};
+const std::vector<int> kFrontRightMotorPorts = {1, 2};
+const std::vector<int> kBackRightMotorPorts = {4, 5};
+const std::vector<int> kBackLeftMotorPorts = {6, 7};
+const std::vector<int> kFrontLeftMotorPorts = {3, 4};
 
 DrivetrainMotors GetDrivetrainMotors() {
   //   return {
@@ -25,4 +25,4 @@ DrivetrainMotors GetDrivetrainMotors() {
 interface::MotorAndEncoder GetShooterMotor() {}
 
 interface::Motor GetTransferMotor() {}
-}  // namespace robot
+}  // namespace constants
